+++ changelog.md
@@ -17,6 +17,7 @@ Added / reworked:
 * <cj>2 player mode</c> received a bunch of fixes, and linking to others is now way clearer
 * Rooms will by default <cg>automatically pin</c> the level played by the host, making it easier for everyone to join
 * Discord linking is now <cg>more interactive</c> and simpler to do
+* Defer Preloading is now enabled by default, which means Globed won't increase load times and instead will load all icons when joining a level while connected
 * Credits menu is now automatically updated server-side
 * UI has been touched up in many places for better looks and clarity
 * <cj>Streamer Mode</c> setting was added, which will hide room password and ID